refactor: move info icons to separate component
